[Rename] to `loggerChannel.ts`
